Moving reducers v
