+++ src/app/components/exui-logger/exui-logger.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit, OnDestroy, Input} from '@angular/core';
 import { NGXLogger } from 'ngx-logger';
-import { LoggerMessage } from 'src/app/services/logger/reducers/logger.reducer';
+import { LoggerMessage } from 'src/app/store/reducers/logger.reducer';
 import { Store, select } from '@ngrx/store';
 import { getDebugMessagesState, getInfoMessagesState, getWarningMessagesState, getErrorMessagesState,
   getFatalMessagesState, getTraceMessagesState } from 'src/app/services/logger/selectors/logger.selector';
